+++ HISTORY.rst
@@ -1,10 +1,12 @@
 History
 =======
 
-0.1.2 (unreleased)
+0.1.4 (2024-10-14)
 ------------------
 
 - Fix bug with file conflict check: #8
+- Dock the plugin to the right side panel, add pagination, search bar for projects: #10
+- Add UI communication system: #11
 
 
 0.1.1 (2024-10-08)